+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file. Be aware th
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.0.20-Release] 2022-10-08
+### Changed
+- Money messages should now appear only in windows you have enabled money display for, instead of always being forced into the primary chat window as before.
+
+### Fixed
+- Added a blacklist for general non-event filters, to avoid normal chat messages in guild- or party chat or similar being interpreted as loot or currency messages and filtered.
+
 ## [1.0.19-Release] 2022-09-24
 ### Fixed
 - Added the missing Evoker color entry to make this work for Dragonflight.